split export.c
