+++ includes/builtins.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   builtins.h                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hrother <hrother@student.42vienna.com>     +#+  +:+       +#+        */
+/*   By: aprevrha <aprevrha@student.42vienna.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/30 18:12:36 by hrother           #+#    #+#             */
-/*   Updated: 2024/04/30 18:14:02 by hrother          ###   ########.fr       */
+/*   Updated: 2024/05/06 22:19:58 by aprevrha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
