+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: aprevrha <aprevrha@student.42vienna.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/05 16:13:22 by hrother           #+#    #+#             */
-/*   Updated: 2024/05/06 21:17:46 by aprevrha         ###   ########.fr       */
+/*   Updated: 2024/05/06 22:17:24 by aprevrha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -131,7 +131,6 @@ int					exec_cmd_list(t_list *cmd_list, t_evars *evars);
 t_list				*lexer(const char *line);
 
 // list_utils.c
-void				destroy_list(t_list *lst);
 void				ft_lstadd_back(t_list **lst, t_list *new);
 void				ft_lstadd_front(t_list **lst, t_list *new);
 void				ft_lstclear(t_list **lst, void (*del)(void *));
